requirements.txt
black==21.12b0
click==8.0.3
cycler==0.11.0
einops==0.3.2
fonttools==4.28.5
kiwisolver==1.3.2
matplotlib==3.5.1
mypy-extensions==0.4.3
numpy==1.21.5
packaging==21.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.1
pyparsing==3.0.6
python-dateutil==2.8.2
six==1.16.0
tomli==1.2.3
torch==1.10.1
torchsummary==1.5.1
torchvision==0.11.2
typing-extensions==4.0.1
